Wrote some login tests and moved over to new generic exception response.
