+++ src/main/java/com/two/workflow_tests/UserLogin.java
@@ -1,0 +1,11 @@
+package com.two.workflow_tests;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserLogin {
+    private final String email;
+    private final String password;
+}
